Redesigned About page with tabs for 'About', 'Thanks' and 'License'.  I think this tidies a page that had become a bit of a dumping ground for information we didn't know what to do with.
The CREDITS file now clearly seperates people by role, the list had grown rather large mainly because we have lots of translators and this is now clear.

The About page no longer advises about plugins that are incompatible as this functionality is now in the plugin manager and never really belongs on a credits page.  The page also no longer mentions features that are disabled, these messages pre-date our config GUI where it's possible to see what's enabled/disabled.